# src/main/resources/openapi/documentation.yaml
openapi: "3.0.3"
info:
  title: "critica_backend API"
  description: "critica_backend API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
  - url: "https://seahorse-app-qs7ri.ondigitalocean.app"
  - url: "https://critika.net"
paths:
...
  /api/user/settings:
    get:
      tags:
        - "User"
      security:
        - AccessToken: [ ]
      description: "Get user settings"
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
  /api/user/settings/change-language:
    post:
      tags:
        - "User"
      security:
        - AccessToken: [ ]
      description: "Change language"
      requestBody:
        content:
          text/plain:
            schema:
              type: "string"
        required: true
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
  /api/user/settings/change-password:
    post:
      tags:
        - "User"
      security:
        - AccessToken: [ ]
      description: "Change password"
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
  /api/user/settings/change-public-visibility:
    post:
      tags:
        - "User"
      security:
        - AccessToken: [ ]
      description: "Change public visibility settings"
      requestBody:
        content:
          application/json:
            schema:
              type: "boolean"
        required: true
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
  /api/user/settings/change-push-notifications:
    post:
      tags:
        - "User"
      security:
        - AccessToken: [ ]
      description: "Change push notification settings"
      requestBody:
        content:
          application/json:
            schema:
              type: "boolean"
        required: true
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
  /api/user/settings/change-username:
    post:
      tags:
        - "User"
      security:
        - AccessToken: [ ]
      description: "Change username"
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
  /api/user/settings/request-promotion:
    post:
      tags:
        - "User"
      security:
        - AccessToken: [ ]
      description: "Request promotion to admin"
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: string
  /api/user/settings/verify-email:
    post:
      tags:
        - "User"
      security:
        - AccessToken: [ ]
      description: "Verify user email"
      requestBody:
        content:
          text/plain:
            schema:
              type: "string"
        required: true
      responses:
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
  /api/user/statistics:
    get:
      tags:
        - "User"
      description: "Get personal rating"
      security:
        - AccessToken: [ ]
      responses:
        "401":
          description: "Unauthorized"
          content:
            application/json:
              schema:
                type: "string"
              examples:
                Example#1:
                  value: "No user id found"
        "404":
          description: "Not Found"
          content:
            application/json:
              schema:
                type: "object"
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResponse"
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    FirebaseToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RoleResponse:
      type: object
      properties:
        id:
          type: "string"
        role:
          type: "string"
        gamesWon:
          type: "integer"
          format: "int32"
        gamesTotal:
          type: "integer"
          format: "int32"
        bonusPoints:
          type: "integer"
          format: "int32"
    RatingResponse:
      type: object
      properties:
        id:
          type: "string"
        userId:
          type: "string"
        totalPoints:
          type: "integer"
          format: "int32"
        bonusPoints:
          type: "integer"
          format: "int32"
        malusPoints:
          type: "integer"
          format: "int32"
        bestMovePoints:
          type: "integer"
          format: "int32"
        roleStatistics:
          type: "array"
          items:
            $ref: "#/components/schemas/RoleResponse"
...
    Error:
      type: "object"
      properties:
        status:
          type: "integer"
          format: "int32"
          example: 400
        message:
          type: "string"
          description: "A description of the error."
          example: "Invalid ID"
      required:
        - "status"
        - "message"
  responses:
    BadRequest:
      description: "Malformed data or otherwise invalid request"
      schema:
        $ref: "#/components/schemas/Error"
    NotFound:
      description: "The requested component was not found"
      schema:
        $ref: "#/components/schemas/Error"
    InternalServerError:
      description: "An internal server error occurred"
      schema:
        $ref: "#/components/schemas/Error"
    Forbidden:
      description: "The user does not have the required permissions to access this resource"
      schema:
        $ref: "#/components/schemas/Error"
    Unauthorized:
      description: "The user is not authenticated"
      schema:
        $ref: "#/components/schemas/Error"